Add logger shim
